+++ web/login.php
@@ -70,11 +70,11 @@ $template = new WebTemplate($c, $login->getUsername(), "Login page &bull; CoLWI"
           </div>
         </div>
         <div class="form-group row">
-          <label class="col-sm-2 form-control-label" for="inputUser">Username</label>
+          <label class="col-sm-2 col-form-label" for="inputUser">Username</label>
           <div class="col-sm-10"><input class="form-control" type="text" id="inputUser" placeholder="Username" name="bah" required autofocus></div>
         </div>
         <div class="form-group row">
-          <label class="col-sm-2 form-control-label" for="inputPass">Password</label>
+          <label class="col-sm-2 col-form-label" for="inputPass">Password</label>
           <div class="col-sm-10"><input class="form-control" type="password" id="inputPass" placeholder="Password" name="dun" required></div>
         </div>
         <div class="row">